Style: 전체 코드 리포맷 & Optimize Imports

+++ src/main/java/com/codeit/HRBank/mapper/BackupMapper.java
@@ -12,17 +12,17 @@ import org.springframework.stereotype.Component;
 @Component
 public class BackupMapper {
 
-    public BackupDto toDto(Backup backup){
+    public BackupDto toDto(Backup backup) {
         // file이 null인 경우에 대한 처리
         Long fileId = (backup.getFile() != null) ? backup.getFile().getId() : null;
 
         return new BackupDto(
-                backup.getId(),
-                backup.getWorker(),
-                backup.getStartedAt(),
-                backup.getEndedAt(),
-                backup.getStatus(),
-                fileId
+            backup.getId(),
+            backup.getWorker(),
+            backup.getStartedAt(),
+            backup.getEndedAt(),
+            backup.getStatus(),
+            fileId
         );
     }
 
@@ -32,8 +32,8 @@ public class BackupMapper {
 
         // 1. Slice<Backup>의 content를 Stream을 이용해 List<BackupDto>로 변환
         List<BackupDto> dtoList = backupSlice.getContent().stream()
-                .map(this::toDto)
-                .collect(Collectors.toList());
+            .map(this::toDto)
+            .collect(Collectors.toList());
 
         // 2. CursorPageResponseBackupDto에 필요한 커서 및 페이지 정보 계산
         Long nextIdAfter = null;
@@ -45,12 +45,12 @@ public class BackupMapper {
 
         // 3. 변환된 DTO 리스트와 커서 정보를 담아 최종 응답 객체 생성
         return new CursorPageResponseBackupDto<>(
-                dtoList,
-                null, // nextCursor는 필요시 추가
-                nextIdAfter,
-                backupSlice.getSize(),
-                (int) backupSlice.getNumberOfElements(),
-                backupSlice.hasNext()
+            dtoList,
+            null, // nextCursor는 필요시 추가
+            nextIdAfter,
+            backupSlice.getSize(),
+            backupSlice.getNumberOfElements(),
+            backupSlice.hasNext()
         );
     }
