+++ src/main/java/com/codeit/HRBank/mapper/DepartmentMapper.java
@@ -13,25 +13,25 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class DepartmentMapper {
 
-  public DepartmentDto toDto(Department department) {
-    return new DepartmentDto(
-        department.getId(),
-        department.getName(),
-        department.getDescription(),
-        department.getEstablishedDate()
-        // department.getEmployeeCount() // 구현 전 컴파일 에러 방지 주석
-    );
-  }
-
+    public DepartmentDto toDto(Department department) {
+        return new DepartmentDto(
+            department.getId(),
+            department.getName(),
+            department.getDescription(),
+            department.getEstablishedDate()
+            // department.getEmployeeCount() // 구현 전 컴파일 에러 방지 주석
+        );
+    }
 
 
     // Slice<Department>을 CursorPageResponseDepartmentDto<DepartmentDto>로 변환하는 메서드
-    public CursorPageResponseDepartmentDto<DepartmentDto> toDtoSlice(Slice<Department> departmentSlice) {
+    public CursorPageResponseDepartmentDto<DepartmentDto> toDtoSlice(
+        Slice<Department> departmentSlice) {
 
         // 1. Slice<Department>의 content를 Stream을 이용해 List<DepartmentDto>로 변환
         List<DepartmentDto> dtoList = departmentSlice.getContent().stream()
-                .map(this::toDto)
-                .collect(Collectors.toList());
+            .map(this::toDto)
+            .collect(Collectors.toList());
 
         // 2. CursorPageResponseDepartmentDto에 필요한 커서 및 페이지 정보 계산
         Long nextIdAfter = null;
@@ -43,12 +43,12 @@ public class DepartmentMapper {
 
         // 3. 변환된 DTO 리스트와 커서 정보를 담아 최종 응답 객체 생성
         return new CursorPageResponseDepartmentDto<>(
-                dtoList,
-                null, // nextCursor는 필요시 추가
-                nextIdAfter,
-                departmentSlice.getSize(),
-                (int) departmentSlice.getNumberOfElements(),
-                departmentSlice.hasNext()
+            dtoList,
+            null, // nextCursor는 필요시 추가
+            nextIdAfter,
+            departmentSlice.getSize(),
+            departmentSlice.getNumberOfElements(),
+            departmentSlice.hasNext()
         );
     }
